simplifying jsx compilation and using a connection pool of workers
