+++ Gemfile
@@ -13,6 +13,7 @@ gem "connection_pool"
 
 gem "turbolinks"
 gem "sass"
+gem "coffee-rails"
 gem "bourbon"
 gem "browserify-rails"
